% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.division_name_map}
\alias{.division_name_map}
\title{Map Division Names}
\format{
An object of class \code{list} of length 13.
}
\usage{
.division_name_map
}
\description{
Map Division Names
}
\keyword{internal}
