% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geofabric_readme.R
\name{geofabric_readme}
\alias{geofabric_readme}
\title{Read Geofabric README}
\usage{
geofabric_readme()
}
\value{
A character vector containing the lines of the README file.
}
\description{
This function returns the contents of the Geofabric README file.
}
\examples{
readme_contents <- geofabric_readme()
cat(readme_contents, sep = "\n")
}
