% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_drainage_divisions.R
\name{list_drainage_divisions}
\alias{list_drainage_divisions}
\title{List Available Drainage Divisions}
\usage{
list_drainage_divisions()
}
\value{
Character vector of available drainage division names
}
\description{
List Available Drainage Divisions
}
