% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_drainage_division.R
\name{load_drainage_division}
\alias{load_drainage_division}
\title{Load Drainage Division Data}
\usage{
load_drainage_division(
  division_name,
  layers = c("AWRADrainageDivision", "RiverRegion", "AHGFNetworkStream", "AHGFWaterbody")
)
}
\arguments{
\item{division_name}{Character; name of the drainage division}

\item{layers}{Character vector; which layers to load ("AWRADrainageDivision", "RiverRegion", 
"AHGFNetworkStream", "AHGFWaterbody"). Default loads all.}
}
\value{
List of sf objects for the specified drainage division
}
\description{
Load Drainage Division Data
}
